Update test descriptions
On naming:
Update and align test descriptions to "unit test" and "ruling test". The terms are still not very clear. Ruling tests could be called simply integration. Unit could be just renamed to lint testing. The term "lint" is already in use in the data file name and throughout runRule.ts.

On code organization:
Currently the 2 existing test.ts files are identical, expect for the Rule file name. In order to reduce code duplications, and unwanted differences like this commit tries to fix, separate test.ts files could be eliminated, and the Rule file names used as input for a parametrized testing setup.